wip - FluentValidation

+++ Application/Logic/Employee/Dtos/EmployeeDto.cs
@@ -10,7 +10,7 @@ namespace Application.Dtos
 
         public string Surname { get; set; }
 
-        public int IdNumber { get; set; }
+        public long IdNumber { get; set; }
 
         public bool IsDeleted { get; set; }
